\name{af}
\alias{af}
\title{Convert some columns of a data.frame to factors}
\description{
  Conveniently convert some columns of data.frame into factors.
}
\usage{
  af(DataFrame, Cols)
}
\arguments{
  \item{DataFrame}{a \code{data.frame}}
  \item{Cols}{column names or indices to be converted}
}
\details{
  It performs conversion of some columns in a \code{data.frame} into factors conveniently. 
}
\value{
  Returns a \code{data.frame} with converted columns.
}
\author{Kyun-Seop Bae k@acr.kr}
