\name{aspirinCHD}
\alias{aspirinCHD}
\docType{data}
\title{
  An example data for meta-analysis - aspirin in coronary heart disease
}
\description{
  The data is from 'Canner PL. An overview of six clinical trials of aspirin in coronary heart disease. Stat Med. 1987'
}
\usage{aspirinCHD}
\format{
  A data frame with 6 rows.
  \describe{
    \item{\code{y1}}{death event count of aspirin group}
    \item{\code{n1}}{total subjet of aspirin group}
    \item{\code{y2}}{death event count of placebo group}
    \item{\code{n2}}{total subject of placebo group}
  }
}
\details{
  This data is for educational purpose.
}
\references{
  Canner PL. An overview of six clinical trials of aspirin in coronary heart disease. Stat Med. 1987;6:255-263.
}
\keyword{datasets}
