\name{BasicUtil}
\alias{pivotJ}
\alias{sortColName}
\alias{sumANOVA}
\alias{sumREG}
\alias{lsm0}
\alias{llsm0}
\alias{bL}
\alias{GrpCode}
\alias{plotDiff}
\alias{plotDunnett}
\alias{CheckAlias}
\title{Internal Functions}
\description{Internal functions}
\details{These are not to be called by the user.}
\keyword{internal}
