\name{DoEdata}
\alias{DoEdata}
\alias{score1}
\alias{bondreturn}
\alias{bulblife}
\alias{simxy1}
\alias{simxy2}
\alias{qcpass}
\alias{oneway}
\alias{edupgm}
\alias{transfat}
\alias{brandtar}
\alias{bacteria}
\alias{temptrend}
\alias{dietwt}
\alias{pigfeed}
\alias{hlpump}
\alias{BP}
\alias{fertpest1}
\alias{fertpest2}
\alias{defect1}
\alias{toollife}
\alias{score2}
\alias{rosemite}
\alias{sodium}
\alias{dampoll1}
\alias{dampoll2}
\alias{twoway}
\alias{dstrength}
\alias{tirewear1}
\alias{tirewear2}
\alias{tirewear3}
\alias{defect2}
\alias{keyspeed}
\alias{dynamite}
\alias{feedmice1}
\alias{feedmice2}
\alias{ccdyield}
\alias{bearing}
\alias{rsdata0}
\alias{ptensile}
\alias{beanyield}
\alias{leprosy}
\alias{feedsheep}
\docType{data}
\title{
  Example Datasets
}
\description{
  Contains data frames to be used for textbooks
}
\details{
  This contains datasets for textbooks.
}
\keyword{internal}

