\name{Kurtosis}
\alias{Kurtosis}
\title{Kurtosis}
\description{
  Kurtosis with a conventional formula.
}
\usage{
  Kurtosis(x)
}
\arguments{
  \item{x}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. 
}
\value{
  Estimate of kurtosis
}
\author{Kyun-Seop Bae k@acr.kr}
