\name{KurtosisSE}
\alias{KurtosisSE}
\title{Standard Error of Kurtosis}
\description{
  Standard error of the estimated kurtosis with a conventional formula.
}
\usage{
  KurtosisSE(y)
}
\arguments{
  \item{y}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. 
}
\value{
  Standard error of the estimated kurtosis
}
\seealso{
  \code{\link{Kurtosis}}
}
\author{Kyun-Seop Bae k@acr.kr}
