\name{Range}
\alias{Range}
\title{Range}
\description{
  The range, maximum - minimum, as a scalar value.
}
\usage{
  Range(y)
}
\arguments{
  \item{y}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. 
}
\value{
  A scalar value of range
}
\author{Kyun-Seop Bae k@acr.kr}
