\name{sasLM-package}
\alias{sasLM-package}
\docType{package}
\title{'SAS' Linear Model}
\description{
  This is a core implementation of 'SAS' procedures for linear models - GLM, REG, and ANOVA. Well-known 'car' package provides type II and type III SS. However, the results for nested and complex designs often different from those of 'SAS.' Different results does not necessarily mean incorrectness. However, many wants the same results to SAS. This package aims to achieve that.
  Reference: Littell RC, Stroup WW, Freund RJ (2002, ISBN:0-471-22174-0).
}
\details{
  This will serve those who want SAS PROC GLM, REG, and ANOVA in R. 
}
\author{Kyun-Seop Bae k@acr.kr}
