\name{SD}
\alias{SD}
\title{Standard Deviation}
\description{
  Standard deviation of sample.
}
\usage{
  SD(y)
}
\arguments{
  \item{y}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. The length of the vector should be larger than 1.
}
\value{
  Sample standard deviation
}
\author{Kyun-Seop Bae k@acr.kr}
