\name{SEM}
\alias{SEM}
\title{Standard Error of the Sample Mean}
\description{
  The estimate of the standard error of the sample mean
}
\usage{
  SEM(x)
}
\arguments{
  \item{x}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. 
}
\value{
  The estimate of the standard error of the sample mean
}
\author{Kyun-Seop Bae k@acr.kr}
