\name{Skewness}
\alias{Skewness}
\title{Skewness}
\description{
  Skewness with a conventional formula.
}
\usage{
  Skewness(y)
}
\arguments{
  \item{y}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. 
}
\value{
  Estimate of skewness
}
\seealso{
  \code{\link{SkewnessSE}}
}
\author{Kyun-Seop Bae k@acr.kr}
