\name{SkewnessSE}
\alias{SkewnessSE}
\title{Standard Error of Skewness}
\description{
  Standard errof of the skewness with a conventional formula.
}
\usage{
  SkewnessSE(x)
}
\arguments{
  \item{x}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. 
}
\value{
  Standard error of the estimated skewness
}
\seealso{
  \code{\link{Skewness}}
}
\author{Kyun-Seop Bae k@acr.kr}
