\name{Median}
\alias{Median}
\title{Median without NA}
\description{
  median without \code{NA} values.
}
\usage{
  Median(x)
}
\arguments{
  \item{x}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. 
}
\value{
  median value
}
\author{Kyun-Seop Bae k@acr.kr}
