\name{N}
\alias{N}
\title{Number of observations}
\description{
  Number of observations excluding \code{NA} values
}
\usage{
  N(y)
}
\arguments{
  \item{y}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. 
}
\value{
  Count of the observation
}
\author{Kyun-Seop Bae k@acr.kr}
