\name{Skewness}
\alias{Skewness}
\title{Skewness}
\description{
  Skewness with a conventional formula.
}
\usage{
  Skewness(x)
}
\arguments{
  \item{x}{a vector of numerics}
}
\details{
  It removes \code{NA} in the input vector. 
}
\value{
  Estimate of skewness
}
\author{Kyun-Seop Bae k@acr.kr}
